feat(movement, model): changed movement of world and added basic model
camera had issues tracking player. ran into gimbal locks and problems with camera trying to stay "up".
work-around was to revert movement back to world as camera was always to be fixed so we did not need full
movement from player.
this allowed for a fixed camera and player, only now the world being affected by physics.

also added basic model and baked texture for debugging.